+++ app/controllers/authenticated/thread.ts
@@ -1,8 +1,11 @@
 import Controller from '@ember/controller';
+import { tracked } from '@glimmer/tracking';
 import { ThreadRouteModel } from 'potber-client/routes/authenticated/thread';
+import { Threads } from 'potber-client/services/api/types';
 
 export default class ThreadController extends Controller {
   declare model: ThreadRouteModel;
+  @tracked cache: Threads.Read | null = null;
 
   queryParams = ['TID', 'page', 'PID', 'lastReadPost', 'scrollToBottom'];
   TID = '';
@@ -12,7 +15,7 @@ export default class ThreadController extends Controller {
   scrollToBottom = '';
 
   get thread() {
-    return this.model.threadResource.value;
+    return this.model.threadResource.value ?? this.cache;
   }
 
   get pageTitle() {