+++ app/routes/authenticated/settings/settings.config.ts
@@ -139,6 +139,17 @@ export const settingsConfig: Record<
     },
   ],
 
+  transitionsOptions: [
+    {
+      label: 'Dynamisch',
+      data: 'dynamic',
+    },
+    {
+      label: 'Statisch',
+      data: 'static',
+    },
+  ],
+
   gesturesOptions: [
     {
       label: 'Alle',