+++ app/services/settings/types.ts
@@ -9,6 +9,7 @@ export interface Settings {
   hideGlobalAndAnnouncementThreads: boolean;
   replaceForumUrls: boolean;
   goToBottomOfThreadPage: boolean;
+  transitions: Transitions;
   gestures: Gestures;
   debug: boolean;
 }
@@ -48,3 +49,10 @@ export enum Gestures {
   onlySidebar,
   all,
 }
+
+export const Transitions = {
+  dynamic: 'dynamic',
+  static: 'static',
+} as const;
+
+export type Transitions = (typeof Transitions)[keyof typeof Transitions];