+++ tests/unit/services/settings-test.ts
@@ -7,6 +7,7 @@ import SettingsService, {
   Settings,
   SidebarLayout,
   Theme,
+  Transitions,
 } from 'potber-client/services/settings';
 import { setupTest } from 'potber-client/tests/helpers';
 import { module, test } from 'qunit';
@@ -22,6 +23,7 @@ const mockedSettings: Settings = {
   darkenReadPosts: false,
   hideGlobalAndAnnouncementThreads: false,
   goToBottomOfThreadPage: true,
+  transitions: Transitions.static,
   gestures: Gestures.none,
   debug: false,
 };